fix: Fix server-side processing of UPDATE_POSITION messages received from the client; new positions being broadcasted to all other clients in the same game

+++ server/src/main/java/edu/brown/cs32/position/Position.java
@@ -1,3 +1,3 @@
 package edu.brown.cs32.position;
 
-public record Position(float x, float y) {}
+public record Position(double x, double y) {}
